requirements.txt
# Common
python-dotenv

# API
fastapi[all]
uvicorn[standard]
requests
python-multipart
python-jose[cryptography]
passlib[bcrypt]

#Third Party Integrations
pymongo[srv]
gspread
./strava
